Updated the format of the new Post Message Events.
