+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/Announcements.java
@@ -9,7 +9,7 @@ import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.scheduler.ScheduledTask;
-import xyz.olivermartin.multichat.bungee.events.MultiChatBroadcastIRCEvent;
+import xyz.olivermartin.multichat.bungee.events.PostBroadcastEvent;
 
 /**
  * Announcements Management
@@ -37,12 +37,12 @@ public class Announcements {
 
 					message = ChatControl.applyChatRules(message, "announcements", "").get();
 
-					// Alert IRC plugins
-					ProxyServer.getInstance().getPluginManager().callEvent(new MultiChatBroadcastIRCEvent("announcement", message));
-
 					for (ProxiedPlayer onlineplayer : ProxyServer.getInstance().getPlayers()) {
 						onlineplayer.sendMessage(TextComponent.fromLegacyText(ChatColor.translateAlternateColorCodes('&',message)));
 					}
+					
+					// Trigger PostBroadcastEvent
+					ProxyServer.getInstance().getPluginManager().callEvent(new PostBroadcastEvent("announcement", message));
 				}
 
 			}, 0, minutes, TimeUnit.MINUTES);
@@ -131,12 +131,12 @@ public class Announcements {
 
 			message = ChatControl.applyChatRules(message, "announcements", "").get();
 
-			// Alert IRC plugins
-			ProxyServer.getInstance().getPluginManager().callEvent(new MultiChatBroadcastIRCEvent("announcement", message));
-
 			for (ProxiedPlayer onlineplayer : ProxyServer.getInstance().getPlayers()) {
 				onlineplayer.sendMessage(TextComponent.fromLegacyText(ChatColor.translateAlternateColorCodes('&',message)));
 			}
+			
+			// Trigger PostBroadcastEvent
+			ProxyServer.getInstance().getPluginManager().callEvent(new PostBroadcastEvent("announcement", message));
 
 		}
 	}
