+++ multichat/src/main/java/xyz/olivermartin/multichat/bungee/ChatStream.java
@@ -12,8 +12,8 @@ import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
-import xyz.olivermartin.multichat.bungee.events.MultiChatBroadcastIRCEvent;
-import xyz.olivermartin.multichat.bungee.events.MultiChatGlobalIRCEvent;
+import xyz.olivermartin.multichat.bungee.events.PostBroadcastEvent;
+import xyz.olivermartin.multichat.bungee.events.PostGlobalChatEvent;
 
 /**
  * Chat Stream
@@ -83,9 +83,6 @@ public class ChatStream {
 	}
 
 	public void sendMessage(ProxiedPlayer sender, String message) {
-		
-		// Alert IRC plugins
-		ProxyServer.getInstance().getPluginManager().callEvent(new MultiChatGlobalIRCEvent(sender, message, format));
 
 		for (ProxiedPlayer receiver : ProxyServer.getInstance().getPlayers()) {
 			if ( (whitelistMembers && members.contains(receiver.getUniqueId())) || (!whitelistMembers && !members.contains(receiver.getUniqueId()))) {
@@ -107,15 +104,15 @@ public class ChatStream {
 				}
 			}
 		}
+		
+		// Trigger PostGlobalChatEvent
+		ProxyServer.getInstance().getPluginManager().callEvent(new PostGlobalChatEvent(sender, message, format));
 
 		ProxyServer.getInstance().getConsole().sendMessage(buildFormatConsole(sender,format,message));
 
 	}
 
 	public void sendMessage(String message) {
-		
-		// Alert IRC plugins
-		ProxyServer.getInstance().getPluginManager().callEvent(new MultiChatBroadcastIRCEvent("cast", message));
 		
 		for (ProxiedPlayer receiver : ProxyServer.getInstance().getPlayers()) {
 			if ( (whitelistMembers && members.contains(receiver.getUniqueId())) || (!whitelistMembers && !members.contains(receiver.getUniqueId()))) {
@@ -128,6 +125,10 @@ public class ChatStream {
 				}
 			}
 		}
+		
+		// Trigger PostBroadcastEvent
+		ProxyServer.getInstance().getPluginManager().callEvent(new PostBroadcastEvent("cast", message));
+		
 		//TODO
 		System.out.println("\033[33m[MultiChat][CHAT]" + message);
